Correction to iEucS calculation for when O are not included
Added the DEM calculation for weightings when only stream layers are included as targets. When only stream layers were included, the O cost calculation was being performed to a single point on the DEM, rather than to the stream network. Did not appear that the cost was being calculated, but an error was not thrown.